pyproject.toml
[tool.poetry]
name = "speechgeneration"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "src"}
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
torchvision = "^0.14.1"
torchaudio = {url = "https://download.pytorch.org/whl/cu116/torchaudio-0.13.1%2Bcu116-cp310-cp310-linux_x86_64.whl"}
pynktrombonegym = {git = "https://github.com/matsuo-group24-PinkTrombone/PynkTromboneGym.git", rev = "master"}
humanfriendly = "^10.0"
typeguard = "^2.13.3"
pytest = "^7.2.0"
torch = "^1.13.1"
librosa = "^0.9.2"
torch-complex = "^0.4.3"
hydra-core = "^1.3.1"
hydra-colorlog = "^1.2.0"
hydra-optuna-sweeper = "^1.2.0"
pyrootutils = "^1.0.4"
rich = "^13.2.0"
tensorboard = "^2.11.2"
torch-tb-profiler = "^0.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-skip-slow = "^0.0.3"
pre-commit = "^2.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = [
    "--color=yes",
    "--durations=0",
    "--strict-markers",
    "--doctest-modules",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
log_cli = "True"
markers = [
    "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]
